fix admin list under status
