+++ src/@types/OceanNode.ts
@@ -91,6 +91,7 @@ export interface OceanNodeStatus {
   platform: any
   uptime?: number // seconds since start
   codeHash?: string
+  allowedAdmins?: string[]
 }
 
 export interface P2PBroadcastResponse {
