+++ src/components/core/utils/statusHandler.ts
@@ -10,6 +10,7 @@ import { existsEnvironmentVariable, getConfiguration } from '../../../utils/inde
 import { ENVIRONMENT_VARIABLES } from '../../../utils/constants.js'
 import { CORE_LOGGER } from '../../../utils/logging/common.js'
 import { OceanNode } from '../../../OceanNode.js'
+const regex: RegExp = /^(0x)?[0-9a-fA-F]{40}$/
 
 export async function status(
   oceanNode: OceanNode,
@@ -27,6 +28,24 @@ export async function status(
   }
   const config = await getConfiguration()
   const { indexer: indexerDatabase } = oceanNode.getDatabase()
+  const validAddresses = []
+  if (config.allowedAdmins) {
+    for (const admin of config.allowedAdmins) {
+      if (regex.test(admin) === true) {
+        validAddresses.push(admin)
+        CORE_LOGGER.log(
+          LOG_LEVELS_STR.LEVEL_ERROR,
+          `Invalid format for ETH address from ALLOWED ADMINS.`
+        )
+      } else {
+        CORE_LOGGER.log(
+          LOG_LEVELS_STR.LEVEL_ERROR,
+          `Invalid format for ETH address from ALLOWED ADMINS.`
+        )
+      }
+    }
+  }
+
   const status: OceanNodeStatus = {
     id: undefined,
     publicKey: undefined,
@@ -51,7 +70,8 @@ export async function status(
       osVersion: os.version(),
       node: process.version
     },
-    codeHash: config.codeHash
+    codeHash: config.codeHash,
+    allowedAdmins: validAddresses.length > 0 ? validAddresses : []
   }
   if (nodeId && nodeId !== undefined) {
     status.id = nodeId
