+++ src/utils/constants.ts
@@ -22,8 +22,7 @@ export const PROTOCOL_COMMANDS = {
   COMPUTE_STOP: 'stopCompute',
   COMPUTE_GET_STATUS: 'getComputeStatus',
   COMPUTE_GET_RESULT: 'getComputeResult',
-  COMPUTE_INITIALIZE: 'initializeCompute',
-  GET_ADMIN_LIST: 'adminList'
+  COMPUTE_INITIALIZE: 'initializeCompute'
 }
 // more visible, keep then close to make sure we always update both
 export const SUPPORTED_PROTOCOL_COMMANDS: string[] = [
@@ -46,8 +45,7 @@ export const SUPPORTED_PROTOCOL_COMMANDS: string[] = [
   PROTOCOL_COMMANDS.COMPUTE_STOP,
   PROTOCOL_COMMANDS.COMPUTE_GET_STATUS,
   PROTOCOL_COMMANDS.COMPUTE_GET_RESULT,
-  PROTOCOL_COMMANDS.COMPUTE_INITIALIZE,
-  PROTOCOL_COMMANDS.GET_ADMIN_LIST
+  PROTOCOL_COMMANDS.COMPUTE_INITIALIZE
 ]
 
 export const MetadataStates = {
